+++ src/containers/Form/ChatWidgets/__tests__/layout.test.js
@@ -7,6 +7,10 @@ import { shallow } from 'enzyme';
 
 import ChatWidgetsForm from '../layout';
 
+const initialValues = {
+  get: () => {}
+};
+
 describe('<ChatWidgetsForm />', () => {
   it('renders create chatWidgets', () => {
     const rendered = shallow(
@@ -28,6 +32,35 @@ describe('<ChatWidgetsForm />', () => {
         iconColor="#000000"
         borderColor="#000000"
         handleSubmit={() => {}}
+        initialValues={initialValues}
+        digitalChannelsAppsFetching={false}
+        digitalChannelsAppIds={[{ value: 'mockValue', label: 'mockLabel' }]}
+      />
+    );
+    expect(rendered).toMatchSnapshot();
+  });
+  it('renders update chatWidgets', () => {
+    const rendered = shallow(
+      <ChatWidgetsForm
+        name="mockName"
+        shared={false}
+        welcome="hero"
+        fontFamily="arial"
+        fontSize="12pt"
+        headerColor="#000000"
+        headerTextIcons="#000000"
+        chatbg="#000000"
+        agentHeader="#000000"
+        agentText="#000000"
+        customerHeader="#000000"
+        customerText="#000000"
+        systemText="#000000"
+        buttonText="#000000"
+        iconColor="#000000"
+        borderColor="#000000"
+        whitelistedUrls={['http://www.example.com']}
+        handleSubmit={() => {}}
+        initialValues={initialValues}
       />
     );
     expect(rendered).toMatchSnapshot();