+++ src/containers/Form/ChatWidgets/__tests__/validation.test.js
@@ -1,12 +1,22 @@
 import { formValidation } from '../validation';
-import { Map } from 'immutable';
+import { Map, fromJS } from 'immutable';
 
 describe('formValidation', () => {
   it('returns proper object when required fields are provided', () => {
     const values = new Map({
       name: 'mockName',
       contactPoint: 'mockContactPoint',
-      appId: 'mockAppId'
+      appId: 'mockAppId',
+      whitelistedUrls: ['http://www.hello.org']
+    });
+    expect(formValidation(values)).toMatchSnapshot();
+  });
+  it('returns proper object with error messages when wrong values are provided', () => {
+    const values = new Map({
+      name: '',
+      contactPoint: '',
+      appId: null,
+      whitelistedUrls: fromJS(['http://www.hello.org/hello_world', 'https://good.one'])
     });
     expect(formValidation(values)).toMatchSnapshot();
   });