+++ src/containers/Form/ChatWidgets/layout.js
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2015-2018 Serenova, LLC. All rights reserved.
+ * Copyright © 2015-2020 Serenova, LLC. All rights reserved.
  */
 
 /**
@@ -20,6 +20,34 @@ import {
   ListField,
   Detail
 } from 'cx-ui-components';
+import styled from 'styled-components';
+import { isSerializedOrigin } from 'serenova-js-utils/strings';
+
+const Error = styled.span`
+  color: red;
+`;
+const Suggestion = styled.span`
+  color: gray;
+  font-style: italic;
+  font-size: 12px;
+  display: block;
+`;
+const BoldSuggestion = styled(Suggestion)`
+  font-weight: bold;
+`;
+const error = (
+  <Fragment>
+    <Error>
+      {`URL entered is not following 'serialized-origin' format from `}
+      <a href="https://tools.ietf.org/html/rfc6454" target="_blank" rel="noopener noreferrer">
+        RFC 6454
+      </a>
+    </Error>
+    <Suggestion>The structure should follow the pattern:</Suggestion>
+    <BoldSuggestion>scheme "://" host [ ":" port ], where scheme is http or https</BoldSuggestion>
+    <Suggestion>Example: http://www.example.com</Suggestion>
+  </Fragment>
+);
 
 export default function ChatWidgetsForm({
   handleSubmit,
@@ -57,18 +85,19 @@ export default function ChatWidgetsForm({
         inputType="text"
         disabled={disabled}
       />
-      {chatWidgetId === 'create'
-        ? <SelectField
-            name="appId"
-            label="App *"
-            options={!digitalChannelsAppsFetching ? digitalChannelsAppIds : undefined}
-            disabled={disabled}
-          />
-        : <Fragment>
-            <Detail label="App" value={app && app.get('name')} />
-            <Detail label="App Id" value={app && app.get('id')} />
-          </Fragment>
-      }
+      {chatWidgetId === 'create' ? (
+        <SelectField
+          name="appId"
+          label="App *"
+          options={!digitalChannelsAppsFetching ? digitalChannelsAppIds : undefined}
+          disabled={disabled}
+        />
+      ) : (
+        <Fragment>
+          <Detail label="App" value={app && app.get('name')} />
+          <Detail label="App Id" value={app && app.get('id')} />
+        </Fragment>
+      )}
 
       <DetailHeader text="Branding" />
 
@@ -78,14 +107,8 @@ export default function ChatWidgetsForm({
         label="Display Style"
         disabled={disabled}
         options={[
-          {
-            label: 'Button',
-            value: 'button'
-          },
-          {
-            label: 'Tab',
-            value: 'tab'
-          }
+          { label: 'Button', value: 'button' },
+          { label: 'Tab', value: 'tab' }
         ]}
         required
       />
@@ -161,20 +184,20 @@ export default function ChatWidgetsForm({
         label="Prechat Capture"
         disabled={disabled}
         options={[
-          {
-            label: 'Name',
-            value: 'name'
-          },
-          {
-            label: 'Email',
-            value: 'email'
-          }
+          { label: 'Name', value: 'name' },
+          { label: 'Email', value: 'email' }
         ]}
         required
       />
 
       <DetailHeader text="Origin Whitelist" />
-      <ListField name="whitelistedUrls" label="Whitelisted URLs" className="chat-widget-whitelisted-urls" />
+      <ListField
+        name="whitelistedUrls"
+        label="Whitelisted URLs"
+        className="chat-widget-whitelisted-urls"
+        inputValidation={isSerializedOrigin}
+        inputError={error}
+      />
     </form>
   );
 }
@@ -185,7 +208,15 @@ ChatWidgetsForm.propTypes = {
   disabled: PropTypes.bool,
   chatWidgetId: PropTypes.string,
   digitalChannelsAppsFetching: PropTypes.bool,
-  digitalChannelsAppIds: PropTypes.array,
+  digitalChannelsAppIds: PropTypes.oneOfType([
+    PropTypes.arrayOf(
+      PropTypes.shape({
+        label: PropTypes.string,
+        value: PropTypes.string.isRequired
+      })
+    ),
+    PropTypes.object
+  ]),
   app: PropTypes.object,
   displayStyleIsButton: PropTypes.bool
 };