+++ src/redux/modules/entities/chatWidgets/__tests__/selectors.test.js
@@ -38,6 +38,30 @@ describe('getDigitalChannelsAppIds', () => {
     });
     expect(getDigitalChannelsAppIds(state)).toMatchSnapshot();
   });
+  it('maps digital channels when at least one is available for assignment', () => {
+    const state = fromJS({
+      Entities: {
+        digitalChannelsApps: {
+          data: [
+            { name: 'mock-name', id: 'mock-app-id', foo: 'bar' },
+            { name: 'mock-name-2', id: 'mock-app-id-2', foo: 'bar' },
+            { name: 'mock-name-3', id: 'mock-app-id-3', foo: 'bar' }
+          ]
+        },
+        chatWidgets: { data: [{ appId: 'mock-app-id' }, { appId: 'mock-app-id-2' }] }
+      }
+    });
+    expect(getDigitalChannelsAppIds(state)).toMatchSnapshot();
+  });
+  it('returns empty map when there is no app available for assignment', () => {
+    const state = fromJS({
+      Entities: {
+        digitalChannelsApps: { data: [{ name: 'mock-name', id: 'mock-app-id', foo: 'bar' }] },
+        chatWidgets: { data: [{ appId: 'mock-app-id' }] }
+      }
+    });
+    expect(getDigitalChannelsAppIds(state)).toMatchSnapshot();
+  });
 });
 
 describe('getDigitalChannelsApp', () => {