+++ src/redux/modules/entities/chatWidgets/selectors.js
@@ -1,8 +1,8 @@
 /*
- * Copyright © 2015-2018 Serenova, LLC. All rights reserved.
+ * Copyright © 2015-2020 Serenova, LLC. All rights reserved.
  */
 
-import { Map } from 'immutable';
+import { Map, List } from 'immutable';
 import { createSelector } from 'reselect';
 import { getSelectedEntity } from '../selectors';
 import { selectFormInitialValues, getCurrentFormValueByFieldName } from '../../form/selectors';
@@ -22,25 +22,23 @@ export const selectChatWidgetFormInitialValues = state => {
   return selectFormInitialValues(state);
 };
 
-const getDigitalChannelsApps = state => state.getIn(['Entities', 'digitalChannelsApps', 'data']);
+const getChatWidgets = state => state.getIn(['Entities', 'chatWidgets', 'data'], new List([]));
+const getDigitalChannelsApps = state => state.getIn(['Entities', 'digitalChannelsApps', 'data'], new List([]));
 
-export const getDigitalChannelsAppIds = createSelector(getDigitalChannelsApps, getDigitalChannelsApps => {
-  return getDigitalChannelsApps !== undefined
-    ? getDigitalChannelsApps
-        .map(app => ({
-          value: app.get('id'),
-          label: app.get('name')
-        }))
-        .toJS()
-    : undefined;
-});
+export const getDigitalChannelsAppIds = createSelector(
+  [getDigitalChannelsApps, getChatWidgets],
+  (digitalChannelsApps, chatWidgets) =>
+    digitalChannelsApps
+      .filter(app => chatWidgets.filter(cw => cw.get('appId') === app.get('id')).size === 0)
+      .map(app => ({ value: app.get('id'), label: app.get('name') }))
+);
 
 export const getDigitalChannelsApp = state => {
-  return getSelectedEntity(state)
-    && getSelectedEntity(state).get('appId')
-    && state.getIn(['Entities', 'digitalChannelsApps', 'data'])
-    && state.getIn(['Entities', 'digitalChannelsApps', 'data']).find(app => app.get('id') === getSelectedEntity(state).get('appId'));
+  return (
+    getSelectedEntity(state) &&
+    getSelectedEntity(state).get('appId') &&
+    getDigitalChannelsApps(state).find(app => app.get('id') === getSelectedEntity(state).get('appId'))
+  );
 };
 
-export const getDisplayStyleIsButton = state =>
-  getCurrentFormValueByFieldName(state, 'displayStyle') === 'button';
+export const getDisplayStyleIsButton = state => getCurrentFormValueByFieldName(state, 'displayStyle') === 'button';